pyproject.toml
[project]
name = "AirflowDBTSnowflake"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "tyson", email = "tysonbarretto1991@gmail.com"},
]
dependencies = ["dbt-core>=1.9.2", "dbt-snowflake>=1.9.0", "astronomer-cosmos>=1.8.2", "apache-airflow>=2.10.4", "apache-airflow-providers-snowflake>=6.0.0"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
